Move build scripts into its own folder; add unittests
