+++ .gitignore
@@ -3,7 +3,8 @@ __enumerating__*/
 .idea
 *.log
 /node_modules/
-build/
+/build/
+!/scripts/build/
 yarn.lock
 .nyc_output/
 coverage.lcov
