+++ .prettierignore
@@ -1,13 +1,14 @@
 *.json
 !/*.json
 !/schemas/*.json
+!/scripts/build/
 
 LICENSE
 .*ignore
 .editorconfig
 /package-lock.json
 /CODE_OF_CONDUCT.md
-build/
+/build/
 coverage/
 .features.json
 types.d.ts
