+++ package.json
@@ -85,7 +85,7 @@
     "lint:fix": "node --loader=ts-node/esm --no-warnings=ExperimentalWarning scripts/fix/index.ts",
     "fix": "npm run format:fix && npm run lint:fix",
     "stats": "node --loader=ts-node/esm --no-warnings=ExperimentalWarning scripts/statistics.ts",
-    "build": "node --loader=ts-node/esm --no-warnings=ExperimentalWarning scripts/release/build.ts",
+    "build": "node --loader=ts-node/esm --no-warnings=ExperimentalWarning scripts/build/index.ts",
     "gentypes": "node --loader=ts-node/esm --no-warnings=ExperimentalWarning scripts/generate-types.ts",
     "release": "node --loader=ts-node/esm --no-warnings=ExperimentalWarning scripts/release/index.ts",
     "remove-redundant-flags": "node --loader=ts-node/esm --no-warnings=ExperimentalWarning scripts/remove-redundant-flags.ts",
